[TASK] Ensures that filters are applied in the order of parameters from query string
It is important especially for OrderFilter. Order of parameters in query string has to decide the order of applying order filter - otherwise results are unexpected.

Moreover this commit removes `parameterNamespace`. This parameter is useless since it is possible to group filters (and join them with logical `OR` operator) using same parameter name for multiple properties, e.g. configuration:

```
 * @T3api\ApiFilter(
 *     SearchFilter::class,
 *     properties={
 *          "firstName": "partial",
 *          "lastName": "partial",,
 *          "middleName": "partial",
 *     },
 *     arguments={
 *          "parameterName": "search",
 *     }
 * )
```