Replaces usage of `Doctrine\DBAL\FetchMode` into PDO library to make some filters working correctly on TYPO3 8.7
Before that commit filters `Distance`, `Containt` and `Search` (in specific cases) did not work correctly for TYPO3 8.7.

Class `Doctrine\DBAL\FetchMode` does not exists in doctrine/dbal 2.5.x (was implemented in 2.7.x) and TYPO3 8.7 does not allow to require higher version of this package.

When support for TYPO3 8.7 will be dropped then this commit could be reverted (and additionally requirement for `doctrine/dbal` should be added to composer.json)