Fixes middlewares order - ensures that t3api request middleware is run after `prepare-tsfe-rendering`
T3api request middleware has to be run after `prepare-tsfe-rendering` to correctly recognize Extbase persistence table mapping.

Before that commit it was not stable. If `adminpanel` extension was activated it was changing the order of middlewares and then t3api was working correctly. If `adminpanel` was not activated then t3api request middleware was before `typo3/cms-frontend/tsfe` and table mapping didn't work.

resolves #8