[BUGFIX] Resolve caching issues on TYPO3 v13
Because of breaking change in core (#102932) following exception was thrown: "Setup array has not been initialized. This happens in cached Frontend scope where full TypoScript is not needed by the system"

Previous solution to change page cache identifier (hash base) was not applied anymore and had to be moved into event listener.

https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/13.0/Breaking-102932-RemovedTypoScriptFrontendControllerHooks.html